Mod: Rename Artist.vue
